finishing react tutorial
